Na values for the data sets
Only the yields have have missing values. The reason there are so many missing values on some month yields are because the bonds simply did not exist (the government was not issuing them at the time), best to just use 3m,6m,and 1yr for short term yields. there is also 1 missing value for all of them. This is not going to end up mattering because there is so much data. When using /working with the yields dataset just add b.dropna to the end