delivrable 3 part 2
Drop dupes is erasing good data. Ex: the 10 year bonds and 6 month bonds hit 3% when going into and coming out of a recession. Its the same value but at different times so it messes a lot of things up especially when putting it in a line plot. That's why I have # before the code. So the it doesn't actually run.